backend/requirements.txt
langchain-groq
langchain
langchain-chroma
fastapi
uvicorn
pydantic
python-dotenv
groq
langchain-community
langchain-core
tiktoken
sentence-transformers